Add files via upload
Repository Name: treasure-island-startDAY3

Description:

This repository contains the initial stage of the "Treasure Island" text-based adventure game, inviting players on an exciting journey to find hidden treasure. "LET'S FIND THIS TREASURE"!

Game Description:

Upon running the Python script, users are greeted with an ASCII art of the Treasure Island and a welcoming message. The game prompts users to make choices at different stages by inputting their decisions using the command line. The choices guide the players through a series of challenges:

The initial choice between 'left' or 'right' at a crossroad.
If 'left' is chosen, the player faces a decision to 'wait' for a boat or 'swim' across a lake to reach an island.
Upon 'waiting' and reaching the island, the player encounters a house with three doors: red, yellow, and blue.
Based on the door color chosen ('red', 'yellow', or 'blue'), different outcomes are revealed.
'Red': Leads to a room full of fire (GAME OVER).
'Yellow': The player finds the treasure and wins the game.
'Blue': Entering a room filled with beasts (GAME OVER).
Other choices result in the game ending.
This repository serves as a starting point for further development or modifications to the game. Future updates could include additional storylines, enhanced user interactions, or graphical user interface implementation.

Feel free to fork this repository and expand upon the existing codebase to create a more immersive and engaging gaming experience! Join the adventure and LET'S FIND THIS TREASURE together!